move location file generator
