+++ beheer/generateLIN-PATfiles/requirements.txt
@@ -1,0 +1,2 @@
+requests==2.32.3
+pandas==2.0.2
