+++ beheer/generateLIN-PATfiles/resources/retrieveLinetypeInfo.rq
@@ -1,0 +1,27 @@
+PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
+PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
+PREFIX owl: <http://www.w3.org/2002/07/owl#>
+PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
+PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
+PREFIX dcterms: <http://purl.org/dc/terms/>
+PREFIX sh: <http://www.w3.org/ns/shacl#>
+
+
+SELECT DISTINCT ?tgtLinetype ?tgtLinetypeName ?tgtLinetypeInfo ?tgtLinetypeID
+WHERE {
+    ?Linetype a owl:Class .
+    ?Linetype skos:prefLabel "Lijntype" .
+    ?tgtLinetype rdfs:subClassOf+ ?Linetype .
+    FILTER NOT EXISTS {
+        ?child rdfs:subClassOf ?tgtLinetype .
+    }
+    
+    ?tgtLinetype skos:prefLabel ?tgtLinetypeName .
+    ?tgtLinetype skos:definition ?tgtLinetypeInfo .
+    
+    OPTIONAL {
+        ?tgtLinetype ?idProperty ?tgtLinetypeID .
+        ?idProperty a owl:AnnotationProperty .
+        ?idProperty skos:prefLabel "ID" .
+    }
+}
