--- code/generateLIN-PATfiles/requirements.txt
@@ -1,2 +1,0 @@
-requests==2.32.3
-pandas==2.0.2
